+++ src/components/routes/Lists.js
@@ -3,14 +3,12 @@ import { Link } from 'react-router-dom'
 import axios from 'axios'
 
 import apiUrl from './../../apiConfig'
+import CheckMark from './CheckMark'
 
 import messages from './../AutoDismissAlert/messages'
 
 const Lists = (props) => {
   const [lists, setLists] = useState([])
-  const [checked, setChecked] = useState({
-    checkmark: []
-  })
 
   const { msgAlert } = props
   console.log(props)
@@ -44,27 +42,7 @@ const Lists = (props) => {
       })
   }, [])
 
-  const checkMark = (id) => setChecked((e) => {
-    console.log('checkmark is currently ', id, !e.checkmark)
-    if (e.checkmark.includes(id)) {
-      e.checkmark.pop(id)
-    } else {
-      e.checkmark.push(id)
-    }
-    return { checkmark: e.checkmark }
-  })
-
-  const listsJsx = lists.map(list => (
-    <li key={list._id}>
-      <div className="list-row">
-        <input className="checkbox" type="checkbox" onChange={() => checkMark(list._id)} value={list.checkmark}/>
-      </div>
-
-      <div className="list-row" style= {{ textDecoration: checked.checkmark.includes(list._id) ? 'line-through' : 'none' }} >
-        <Link to={`/lists/${list._id}`}>{list.name}</Link>
-      </div>
-    </li>
-  ))
+  const listsJsx = lists.map(list => <CheckMark key={list._id} list={list} />)
 
   return (
     <div className="list-style">
